edge_simulator/receive_msg_module/requirements.txt
asyncio==3.4.3
azure-iot-device==2.11.0
structlog==21.5.0